Feat: add RegisterParent component to UserRegistration Form
that fetches all registered students from the backend in an async function and lists them in a dropdown menu (for now). When there are more students, a search bar will need to be implemented. The selection of the student and addition to formState is not yet completed, and backend api route is not yet working (currently 404 not found when trying to register a parent user).